+++ lib/rumbleship-beeline.d.ts
@@ -1,8 +1,9 @@
-import { HoneycombSpan, HoneycombConfiguration } from './honeycomb.interfaces';
+import { HoneycombSpan, HoneycombConfiguration, IAsyncTracker } from './honeycomb.interfaces';
 export declare class RumbleshipBeeline {
     private context_id;
     private static beeline;
-    static FinishersByContextId: Map<string, () => any>;
+    static TrackedContextbyContextId: Map<string, any>;
+    static HnyTracker: IAsyncTracker;
     private static initialized;
     /**
      * @param configureBeeline `require('honeycomb-beeline')`
@@ -38,18 +39,22 @@ export declare class RumbleshipBeeline {
     withSpan<T>(metadataContext: object, fn: (span: HoneycombSpan) => T, rollupKey?: string): T;
     withAsyncSpan<T>(this: RumbleshipBeeline, metadata_context: object, fn: (span: HoneycombSpan) => Promise<T> | T): Promise<T>;
     withTrace<T>(metadataContext: object, fn: () => T, withTraceId?: string, withParentSpanId?: string, withDataset?: string): T;
-    finishRumbleshipContextTrace(): void;
     startTrace(span_data: object, traceId?: string, parentSpanId?: string, dataset?: string): HoneycombSpan;
     finishTrace(span: HoneycombSpan): void;
     startSpan(metadataContext: object, spanId?: string, parentId?: string): HoneycombSpan;
     finishSpan(span: HoneycombSpan, rollup?: string): void;
     startAsyncSpan<T>(metadataContext: object, fn: (span: HoneycombSpan) => T): T;
-    bindFunctionToTrace<T>(fn: () => T): T;
+    bindFunctionToTrace<T>(fn: () => T): () => T;
     runWithoutTrace<T>(fn: () => T): T;
     addContext(context: object): void;
     removeContext(context: object): void;
     marshalTraceContext(context: HoneycombSpan): string;
-    unmarshalTraceContext(context_string: string): HoneycombSpan;
+    /**
+     *
+     * @param context_string The wrapped beeline expects a string, even if it is empty. We accept
+     * undefined because that's more typesafe and cast to the empty string.
+     */
+    unmarshalTraceContext(context_string?: string): HoneycombSpan | object;
     getTraceContext(): HoneycombSpan;
     traceActive(): boolean;
 }