+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file. Starting wi
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Added
+  * Enable RumbleshipContext-specific tracking by directly interfacing with native Honeycomb Tracker to enable
+### Removed
+  * Map of finishers + extra fn to use it, in favor just using HnyTracker directly
+### Changed
+  * Expose RumbleshipBeeline.FinishersByContextId
+  * unmarshalTraceContext() accepts `string|undefined`; casts to `''` when passing to wrapped beeline
+### Fixed
+  * Typings on `bindFunctionToTrace()`
+### Deprecated
+### Security
+
+
 ## [2.0.1] -- 2020-04-28
 
 ### Changed